fixing buggy version of pytest on mac/darwin, refactoring code out of init adding task, adding tests
