+++ jupyanno/__init__.py
@@ -1,232 +1,4 @@
-import base64
-import inspect
-import json
-import os
-from collections import namedtuple
-from io import BytesIO
-from itertools import cycle
-
-import matplotlib.pyplot as plt
-import numpy as np
-import pandas as pd
-import seaborn as sns
-from IPython.display import display, Javascript
-from PIL import Image
-from scipy.stats import binom
-from six.moves.urllib.parse import urlparse, parse_qs
-
-TaskData = namedtuple('TaskData',
-                      ['task', 'data_df', 'label_col', 'image_key_col',
-                       'base_img_dir', 'base_sheet_url', 'sheet_id'])
-
-
-def setup_appmode():
-    js_str = """$('#appmode-leave').hide()
-        // Hides the edit app button.
-        $('#appmode-busy').hide()
-        // Hides the kernel busy indicator.
-        IPython.OutputArea.prototype._should_scroll = function(lines) {
-            return false
-            // disable scrolling
-        }"""
-    sns.set_style("whitegrid", {'axes.grid': False})
-    display(Javascript(js_str))
-
-
-def get_app_user_id():
-    """
-    Get the userid from the `jupyter_notebook_url`
-    injected by the appmode extension (if in use)
-    otherwise return a 'nobody'
-    :return: appmode username of current user
-    >>> get_app_user_id()
-    'nobody'
-    >>> jupyter_notebook_url = 'https://a.b.c?user=dan#hello'
-    >>> get_app_user_id()
-    'dan'
-    """
-    cur_url = globals().get('jupyter_notebook_url', None)
-    if cur_url is None:
-        # black magic to get the 'injected' variable
-        frame = inspect.currentframe()
-        try:
-            out_locals = frame.f_back.f_locals
-            cur_url = out_locals.get('jupyter_notebook_url', 'nobody')
-        finally:
-            del frame
-    qs_info = parse_qs(urlparse(cur_url).query)
-    return qs_info.get('user', ['nobody'])[0]
-
-
-def safe_json_load(in_str):
-    """
-    safely load json strings as dictionaries
-    :param in_str: string with encoded json
-    :return: dict
-    >>> safe_json_load('{"bob": 5}')
-    {'bob': 5}
-    >>> safe_json_load('{"bob": [1,2,3]}')
-    {'bob': [1, 2, 3]}
-    >>> safe_json_load('{"bob": [1,2,3}')
-    Invalid json row Expecting ',' delimiter: line 1 column 15 (char 14)
-    {}
-    """
-    try:
-        return json.loads(in_str)
-    except Exception as e:
-        print('Invalid json row {}'.format(e))
-        return dict()
-
-
-def read_annotation(raw_df):
-    annot_df = raw_df.copy()
-    annot_df['Timestamp'] = pd.to_datetime(annot_df['Timestamp'])
-    annot_df['viewing_time'] = annot_df['viewing_info'].map(
-        lambda x: safe_json_load(x).get('viewing_time', 0))
-    annot_df['viewing_info_dict'] = annot_df['viewing_info'].map(
-        lambda x: safe_json_load(x))
-    annot_df['annotator_class'] = annot_df['annotator'].map(
-        lambda x: x.split('_')[0])
-    annot_df['annotator_name'] = annot_df['annotator'].map(
-        lambda x: ' '.join(x.split('_')[1:]) if x.find('_') else x)
-    annot_df['answer_negativity'] = annot_df['label'].map(
-        lambda x: ('No ' in x) or x == 'No')
-    annot_df['time'] = annot_df['time'].map(float)
-    print('Found', annot_df.shape[0], 'completed annotations')
-    return annot_df
-
-
-def binary_correct(c_row, label_col):
-    """
-    determine if the row is correct or not
-    :param c_row: row from a dataframe (dict-like)
-    :param label_col: the colum where the label information resides
-    :return: boolean if the value is correct or not
-    >>> test_row = {'task': 'Pneumonia', 'value': 'Pneumonia', 'label': 'Pneumonia'}
-    >>> binary_correct(test_row, 'value')
-    True
-    >>> test_row['value'] = 'Influenza'
-    >>> binary_correct(test_row, 'value')
-    False
-    >>> test_row = {'task': 'Influenza', 'value': 'Pneumonia', 'label': None}
-    >>> binary_correct(test_row, 'value')
-    True
-    """
-    if c_row['label'] == c_row[label_col]:
-        return True
-    elif c_row['label'] is None:
-        if c_row[label_col] == c_row['task']:
-            # definitely wrong
-            return False
-        else:
-            # definitely right
-            return True
-    else:
-        # if the label is positive but not what we picked
-        return False
-
-    return None  # we arent sure if it is right or not
-
-
-def read_task_file(in_path):
-    with open(in_path, 'r') as f:
-        annotation_task = json.load(f)
-        data_df = pd.DataFrame(annotation_task['dataset']['dataframe'])
-        label_col = annotation_task['dataset']['output_labels']
-        image_key_col = annotation_task['dataset']['image_path']
-        base_img_dir = annotation_task['dataset']['base_image_directory']
-        base_img_dir = os.path.join(os.path.dirname(in_path), base_img_dir)
-        base_sheet_url = annotation_task['google_forms']['sheet_url']
-        sheet_id = \
-            base_sheet_url.strip('?usp=sharing').strip('/edit').split('/')[-1]
-        return TaskData(annotation_task, data_df, label_col, image_key_col,
-                        base_img_dir, base_sheet_url, sheet_id)
-
-
-def show_my_result(name_list, correct_list, num_questions=30, ax1=None):
-    n_correct = np.arange(num_questions + 1)
-    if ax1 is None:
-        fig, ax1 = plt.subplots(1, 1, figsize=(10, 5), dpi=250)
-    binom_pmf = binom.pmf(n_correct, num_questions, 0.5)
-    binom_cdf = np.cumsum(binom_pmf)
-    ax1.bar(n_correct, binom_pmf / np.max(binom_pmf), color='k', alpha=0.5)
-
-    ax1.plot(n_correct, binom_cdf, 'k-', label='Cumulative')
-    prop_cycle = plt.rcParams['axes.prop_cycle']
-    color_cycle = cycle(prop_cycle.by_key()['color'])
-    y_pos_list = np.linspace(0.2, 0.8, len(name_list))
-    for y_pos, name, correct, color in zip(y_pos_list, name_list, correct_list,
-                                           color_cycle):
-        ax1.axvline(correct, label='{}\nTop {:2.1f}%'.format(
-            name, 100 * binom_cdf[correct]), color=color)
-        ax1.text(correct, y_pos, name)
-    ax1.legend()
-    return ax1
-
-
-def _wrap_uri(data_uri): return "data:image/png;base64,{0}".format(data_uri)
-
-
-def raw_html_render(temp_df):
-    """
-    For rendering html tables which contain HTML information and
-    shouldn't be escaped or cropped
-    :param temp_df:
-    :return:
-    >>> f=pd.DataFrame({'k': [1], 'b': ['<a>']})
-    >>> print(f.to_html())
-    <table border="1" class="dataframe">
-      <thead>
-        <tr style="text-align: right;">
-          <th></th>
-          <th>k</th>
-          <th>b</th>
-        </tr>
-      </thead>
-      <tbody>
-        <tr>
-          <th>0</th>
-          <td>1</td>
-          <td>&lt;a&gt;</td>
-        </tr>
-      </tbody>
-    </table>
-    >>> print(raw_html_render(f))
-    <table border="1" class="dataframe table table-striped table-hover">
-      <thead>
-        <tr style="text-align: right;">
-          <th>k</th>
-          <th>b</th>
-        </tr>
-      </thead>
-      <tbody>
-        <tr>
-          <td>1</td>
-          <td><a></td>
-        </tr>
-      </tbody>
-    </table>
-    """
-    old_wid = pd.get_option('display.max_colwidth')
-    pd.set_option('display.max_colwidth', -1)
-    tab_html = temp_df.to_html(classes="table table-striped table-hover",
-                               escape=False,
-                               float_format=lambda x: '%2.2f' % x,
-                               na_rep='',
-                               index=False,
-                               max_rows=None,
-                               max_cols=None)
-
-    pd.set_option('display.max_colwidth', old_wid)
-    return tab_html
-
-
-def path_to_img(in_path):
-    c_img_data = Image.open(in_path)
-    c_img_data = c_img_data.convert('RGB')
-    out_img_data = BytesIO()
-    c_img_data.save(out_img_data, format='png')
-    out_img_data.seek(0)  # rewind
-    uri = _wrap_uri(base64.b64encode(out_img_data.read()
-                                     ).decode("ascii").replace("\n", ""))
-    return '<img src="{uri}"/>'.format(uri=uri)
+from .utils import get_app_user_id, setup_appmode
+from .task import read_annotation, read_task_file
+from . import widgets
+from . import sheets
